+++ NackEngine/core/physics/bounding/AABBox.cs
@@ -48,21 +48,21 @@ namespace NackEngine.core.physics.bounding
             Point rayOrigin = ray.Origin();
             NVector rayDirection = ray.Direction();
 
-            float tMin = rayT.Min();
-            float tMax = rayT.Max();
+            double tMin = rayT.Min();
+            double tMax = rayT.Max();
 
             for (int ax = 0; ax < 3; ax++)
             {
                 Axis axis = (Axis)ax;
                 Range rangeAxis = GetRangeAxis(axis);
 
-                float invD = 1.0f / rayDirection.GetComponent(axis);
-                float origin = rayOrigin.GetComponent(axis);
+                double invD = 1.0 / rayDirection.GetComponent(axis);
+                double origin = rayOrigin.GetComponent(axis);
 
-                float t0 = (rangeAxis.Min() - origin) * invD;
-                float t1 = (rangeAxis.Max() - origin) * invD;
+                double t0 = (rangeAxis.Min() - origin) * invD;
+                double t1 = (rangeAxis.Max() - origin) * invD;
 
-                if (invD < 0.0f)
+                if (invD < 0.0)
                 {
                     (t0, t1) = (t1, t0);
                 }
@@ -78,14 +78,14 @@ namespace NackEngine.core.physics.bounding
             return true;
         }
 
-        public float Area() {
-            float dx = X.Size();
-            float dy = Y.Size();
-            float dz = Z.Size();
+        public double Area() {
+            double dx = X.Size();
+            double dy = Y.Size();
+            double dz = Z.Size();
 
-            if (dx < 0 || dy < 0 || dz < 0) { return 0.0f; }
+            if (dx < 0 || dy < 0 || dz < 0) { return 0.0; }
 
-            return 2.0f * (dx * dy + dy * dz + dz * dx);
+            return 2.0 * (dx * dy + dy * dz + dz * dx);
         }
 
         public int LongestAxis()
@@ -102,7 +102,7 @@ namespace NackEngine.core.physics.bounding
 
         private void MinPadding()
         {
-            float d = 1e-4f;
+            double d = 1e-4;
             if (X.Size() < d) X = X.Expand(d);
             if (Y.Size() < d) Y = Y.Expand(d);
             if (Z.Size() < d) Z = Z.Expand(d);
